Make git test linux only.
